Add files via upload
Once you have set up the project, the index.php file contains the code for executing a SELECT query and displaying the retrieved data.

You can modify the SELECT statement in the index.php file to fetch specific data from your own database. Customize the query based on your table structure and requirements.

Contributing
Contributions to enhance the functionality or improve the code are welcome. If you have any suggestions or find any issues, feel free to open an issue or submit a pull request.

License
This project is licensed under the MIT License. Feel free to use and modify the code as per the license terms.

Acknowledgments
Special thanks to the PHP and MySQL communities for their excellent documentation and resources that helped in the development of this project.

Contact
If you have any questions or want to reach out regarding this project, you can contact me at [genius.rio121@gmail.com].

Happy coding!